http://bugs.osportfolio.org/jira/browse/OSP-1638
JSF converts spaces into + in making the url for the reflection resource url.  But the resource manager doesn't convert + back to spaces.
Added a url getter that changes spaces into %20 thus bypassing the resource manager's url handler.
I don't know if the resource manager is like that for a reason or what.

git-svn-id: https://source.sakaiproject.org/svn/osp/trunk@13512 66ffb92e-73f9-0310-93c1-f5514f145a0a