http://bugs.sakaiproject.org/jira/browse/SAK-6106
removed the owner and publisher from a scaffolding before xmlencoding.  the Agent wrapper had a loop in it thus causing a StackOverflowError.

It seems that the xmlencoder recreates the classes it tries to encode but when it creates the list of properties it seems to include "new" in that list.

git-svn-id: https://source.sakaiproject.org/svn/osp/trunk@14235 66ffb92e-73f9-0310-93c1-f5514f145a0a