http://bugs.sakaiproject.org/jira/browse/SAK-6855
added the reflection to the push into the security advisor

To Test:
have a wizard with a second user set up as the evaluator.  Fill in the reflection for a wizard as a third user.  then log in as the second user (evaluator) and go into the third user's completed wizard.  Click on the third user's reflection.  Before it would give an http status 500 error.  Now it should show the reflection filled out form.


http://bugs.sakaiproject.org/jira/browse/SAK-6096

when importing a wizard the files for the wizard should now be in a folder called "Imported Wizards files" inside a folder of the name of the wizard.
this folder should be created automatically.



Added some javadoc to the wizard manager

git-svn-id: https://source.sakaiproject.org/svn/osp/trunk@17171 66ffb92e-73f9-0310-93c1-f5514f145a0a