svn merge -c 17148  https://source.sakaiproject.org/svn/osp/trunk U    matrix/tool/src/java/org/theospi/portfolio/matrix/control/SequentialWizardPageController.java U    matrix/tool/src/java/org/theospi/portfolio/matrix/control/WizardPageController.java U    matrix/tool/src/java/org/theospi/portfolio/matrix/control/ListEvaluationItemController.java /Develop/2-3-x/osp$ svn log -r 17148  https://source.sakaiproject.org/svn/osp/trunk ------------------------------------------------------------------------ r17148 | andersjb@iupui.edu | 2006-10-16 17:01:10 -0400 (Mon, 16 Oct 2006) | 7 lines
http://bugs.sakaiproject.org/jira/browse/SAK-6856

removed a session variable that was being stored and not letting the requested page show.

To test:
have a published wizard with two pages.  log in as a viewer of the wizard and fill in the two pages. submit the two pages.  log in as the evaluator for the wizard/pages.  click the first page to evaluate, then go back to the evaluations page by clicking the button at the bottom of the page.  then click on the second page.  now the second page should show up.  When it didn't work, clicking on the second page would bring up the first page
also commented a few things
------------------------------------------------------------------------



git-svn-id: https://source.sakaiproject.org/svn/osp/branches/sakai_2-3-x@17216 66ffb92e-73f9-0310-93c1-f5514f145a0a