svn merge -r 17169:17171 https://source.sakaiproject.org/svn/osp/trunk U    wizard/component/src/webapp/WEB-INF/components.xml U    wizard/api-impl/src/java/org/theospi/portfolio/wizard/mgt/impl/WizardManagerImpl.java U    wizard/api/src/java/org/theospi/portfolio/wizard/mgt/WizardManager.java U    presentation/tool/src/java/org/theospi/portfolio/presentation/tool/DecoratedPage.java /Develop/2-3-x/osp$ svn log -r 17170:17171 https://source.sakaiproject.org/svn/osp/trunk ------------------------------------------------------------------------ r17170 | chmaurer@iupui.edu | 2006-10-17 16:32:02 -0400 (Tue, 17 Oct 2006) | 2 lines
http://bugs.sakaiproject.org/jira/browse/SAK-6864
When changing a layout, we need to clear all of the page's region information.  This was happening correctly, but there was a spot where it was being cleared at the wrong time. 
------------------------------------------------------------------------
r17171 | andersjb@iupui.edu | 2006-10-17 16:55:47 -0400 (Tue, 17 Oct 2006) | 16 lines

http://bugs.sakaiproject.org/jira/browse/SAK-6855

added the reflection to the push into the security advisor

To Test:
have a wizard with a second user set up as the evaluator.  Fill in the reflection for a wizard as a third user.  then log in as the second user (evaluator) and go into the third user's completed wizard.  Click on the third user's reflection.  Before it would give an http status 500 error.  Now it should show the reflection filled out form.


http://bugs.sakaiproject.org/jira/browse/SAK-6096

when importing a wizard the files for the wizard should now be in a folder called "Imported Wizards files" inside a folder of the name of the wizard.
this folder should be created automatically.



Added some javadoc to the wizard manager
------------------------------------------------------------------------



git-svn-id: https://source.sakaiproject.org/svn/osp/branches/sakai_2-3-x@17217 66ffb92e-73f9-0310-93c1-f5514f145a0a