http://bugs.sakaiproject.org/jira/browse/SAK-6854
When clicking the cancel button on the submit cell confirmation page and the cell is a sequetial wizard page then the code will revert the current page index back to the original page.

The cancel button was taking the user to the following page like the confirm button.  This was causing confusion as ppl thought the page after cancel was the page being confirmed

To Test:
have a seq wizard with two pages.  As a user filling out the wizard fill in the first page then click the submit to confirm button.  on the confirm submit page click cancel.  the resulting page should still be the page you were submitting.  the click the submit button again and on the confirmation page confirm the submit.  the resulting page should be the second page of the seq wizard.

git-svn-id: https://source.sakaiproject.org/svn/osp/trunk@17862 66ffb92e-73f9-0310-93c1-f5514f145a0a