svn merge -r 19417:19425 https://source.sakaiproject.org/svn/osp/trunk U    reports/components/src/webapp/WEB-INF/components.xml U    reports/api-impl/project.xml U    reports/api-impl/src/java/org/theospi/portfolio/reports/model/impl/ReportsManagerImpl.java U    reports/api-impl/src/java/org/theospi/portfolio/reports/model/impl/LoadArtifactResultProcessor.java U    reports/api/src/java/org/theospi/portfolio/reports/model/ReportDefinition.java U    reports/api/src/java/org/theospi/portfolio/reports/model/ReportsManager.java U    reports/tool/src/webapp/WEB-INF/reports-config.xml U    reports/tool/src/webapp/WEB-INF/reports-definition.xml U    reports/tool/src/webapp/WEB-INF/reportxsls/matrixStatusByLevel.1.xsl A    reports/tool/src/webapp/WEB-INF/reportxsls/matrixStatusByLevelDDB.1.xsl U    presentation/tool/src/java/org/theospi/portfolio/presentation/control/ViewPresentationControl.java U    presentation/tool/src/webapp/WEB-INF/jsp/presentation/comments.inc /Develop/sakai_2-3-x/osp$ svn log -r 19418:19425 https://source.sakaiproject.org/svn/osp/trunk ------------------------------------------------------------------------ r19418 | bkirschn@umich.edu | 2006-12-12 10:33:21 -0500 (Tue, 12 Dec 2006) | 2 lines
SAK-6947 -- check permissions before allowing addition of comments to presentation/portfolio

------------------------------------------------------------------------
r19425 | andersjb@iupui.edu | 2006-12-12 13:50:13 -0500 (Tue, 12 Dec 2006) | 16 lines

http://bugs.sakaiproject.org/jira/browse/SAK-7359

Added a flag to the ReportDefinition to define if the report is to use the data warehouse or the sakai data source
Added a bunch of new javadoc developer documentation to the Report Manager.
Added javadoc is to be inherited for functions in the Impl that implement the ReportManager interface
Added a function to the interface to get the datasource from reports that are switched on whether or not the parameter is to use the data warehouse dataSource
Added a project dependency in the reports api-impl on the sakai-components-api
Changed the Load ArtifactResultProcessor to pull its datasource from the ReportManagerImpl without the warehouse flag
Documented how the datasources are selected in the impl
Added a sakai dataSource to the Impl
When getting the reports from the Impl it checks if the report uses the warehouse or not against the sakai.properties property "osp.reports.useWarehouse"
Reemoved the dataSource from the ResourceLoader Processor
Commented out the dataSource for the ReportManagerImpl.  If it's filled in the warehouse reports will use that.  if it's not set then the code will try to get the data warehouse dataSource.  If the datawarehouse isn't deployed then the warehouse reports will use the sakai dataSource.  This is where the flag is handy to be able to turn off the data warehouse dependent reports
Added the property for the sakai dataSource in the ReportManagerImpl
Created a direct database report as an example and the associated xsl
Changed a sample report to make it work
------------------------------------------------------------------------



git-svn-id: https://source.sakaiproject.org/svn/osp/branches/sakai_2-3-x@19430 66ffb92e-73f9-0310-93c1-f5514f145a0a