SAK-13520
Checking in raw fix to use joins.  This adds one method to MatrixManager to
check for used scaffolding cells.  It also adds three methods to the
implementation that may be desirable at the API for getting raw counts of the
various kinds of interaction items.  Functional testing checks out for the same
behavior.



git-svn-id: https://source.sakaiproject.org/svn/osp/trunk@46478 66ffb92e-73f9-0310-93c1-f5514f145a0a